[Maintenance] Simplify startup process #6

+++ app/bin/run-bridge.sh
@@ -2,10 +2,22 @@
 
 echo "Install bridge"
 
-cd /roon
+ROON_DIR="/roon-dir"
+
+# might already exist
+if [[ ! -d "$ROON_DIR" ]]; then
+    mkdir $ROON_DIR
+else
+    echo "$ROON_DIR already exists"
+fi
+
+echo "Removing anything in $ROON_DIR ..."
+rm $ROON_DIR/* -Rf
+echo "Removal complete"
+
+cd $ROON_DIR
 
 if [[ ! -d "RoonBridge" ]]; then
-    URL_DIR="https://download.roonlabs.net/builds/"
     declare -A file_dict
     file_dict[x86_64]=RoonBridge_linuxx64.tar.bz2
     file_dict[armv7l]=RoonBridge_linuxarmv7hf.tar.bz2
@@ -15,8 +27,21 @@ if [[ ! -d "RoonBridge" ]]; then
         ARCH="${FORCE_ARCH}"
     fi
     FILENAME=${file_dict["${ARCH}"]}
-    FILE="/roon/$FILENAME"
-    curl -O "$URL_DIR/$FILENAME"
+    echo "We need file [$FILENAME] ..."
+    if [ -f "/files/$FILENAME" ]; then
+        echo "File [$FILENAME] is already available via /files mountpoint"
+        cp /files/$FILENAME $ROON_DIR/
+    else
+        URL_DIR="https://download.roonlabs.net/builds/"
+        if [[ -n "${BASE_URL}" ]]; then
+            echo "Using custom BASE_URL = [${BASE_URL}] ..."
+            URL_DIR="${BASE_URL}"
+        fi
+        echo "About to download [$URL_DIR/$FILENAME] ..."
+        curl -O "$URL_DIR/$FILENAME"
+        echo "Download complete"
+    fi
+    FILE="$ROON_DIR/$FILENAME"
     tar -xf $FILE
     echo "Installation terminated"
 else
@@ -25,6 +50,6 @@ fi
 
 echo "Start bridge"
 
-cd /roon/RoonBridge
+cd $ROON_DIR/RoonBridge
 
 ./start.sh
